SIdeBar 컴포넌트 수정 - github, protfolio 연결, category 컴포넌트로 따로 뺌
근데 밑에 먹어버렸어..
